修复偶尔用户权限判断错误问题 修改微博Emoji存在的BUG

+++ src/main/java/com/lxinet/jeesns/core/service/impl/ArchiveServiceImpl.java
@@ -100,7 +100,7 @@ public class ArchiveServiceImpl implements IArchiveService {
         findArchive.setKeywords(archive.getKeywords());
         //普通会员
         if (member.getIsAdmin() == 0) {
-            if (member.getId() != findArchive.getMemberId()) {
+            if (member.getId().intValue() != findArchive.getMemberId().intValue()) {
                 return false;
             }
         } else {
